+++ BlueprintUICommonControls/Sources/Image.swift
@@ -15,15 +15,21 @@ public struct Image: Element {
     /// not precisely match the size that the element is assigned.
     public var contentMode: ContentMode
 
+    /// iOS 14 added support for Image Descriptions using VoiceOver. This is not always appropriate.
+    /// Set this to `true` to prevent VoiceOver from discribing the displayed image.
+    public var blockAccessibilityDescription: Bool
+
     /// Initializes an image element with the given `UIImage` instance.
     public init(
         image: UIImage?,
         contentMode: ContentMode = .aspectFill,
-        tintColor: UIColor? = nil
+        tintColor: UIColor? = nil,
+        blockAccessibilityDescription: Bool = false
     ) {
         self.image = image
         self.contentMode = contentMode
         self.tintColor = tintColor
+        self.blockAccessibilityDescription = blockAccessibilityDescription
     }
 
     public var content: ElementContent {
@@ -37,6 +43,10 @@ public struct Image: Element {
             config[\.contentMode] = contentMode.uiViewContentMode
             config[\.layer.minificationFilter] = .trilinear
             config[\.tintColor] = tintColor
+            if blockAccessibilityDescription {
+                // Seting `isAccessibilityElement = false` isn't enough here, VoiceOver is very aggressive in finding images to discribe. We need to explicitly remove the `.image` trait.
+                config[\.accessibilityTraits] = UIAccessibilityTraits.none
+            }
         }
     }